+++ cnccode/data_classes/DBEUser.inc.php
@@ -59,6 +59,7 @@ class DBEUser extends DBEntity
     const encryptedCounty = "encryptedCounty";
     const encryptedPostcode = "encryptedPostcode";
     const staffAppraiserFlag = 'staffAppraiserFlag';
+    const passwordLevel = 'passwordLevel';
 
     /**
      * calls constructor()
@@ -332,7 +333,13 @@ class DBEUser extends DBEntity
             self::staffAppraiserFlag,
             DA_YN,
             DA_NOT_NULL
-        ); // does user get overtime in weekdays
+        );
+
+        $this->addColumn(
+            self::passwordLevel,
+            DA_INTEGER,
+            DA_NOT_NULL
+        );
 
         $this->setPK(0);
         $this->setAddColumnsOff();